+++ wicket-common/Cargo.toml
@@ -26,12 +26,12 @@ sha2.workspace = true
 sled-agent-types.workspace = true
 sled-hardware-types.workspace = true
 slog.workspace = true
-slog-error-chain.workspace = true
 thiserror.workspace = true
 tokio.workspace = true
 transceiver-controller.workspace = true
 tufaceous-artifact.workspace = true
 update-engine.workspace = true
+wicketd-commission-types.workspace = true
 
 [dev-dependencies]
 toml.workspace = true
