+++ wicketd-commission-types/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "wicketd-commission-types"
+version = "0.1.0"
+edition.workspace = true
+license = "MPL-2.0"
+
+[lints]
+workspace = true
+
+[dependencies]
+wicketd-commission-types-versions.workspace = true
+omicron-workspace-hack.workspace = true
