+++ wicketd-commission-types/versions/src/initial/mod.rs
@@ -1,0 +1,7 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
+//! Version `INITIAL` of the wicketd commissioning API.
+
+pub mod rack_setup;
