+++ wicketd/Cargo.toml
@@ -57,6 +57,7 @@ uuid.workspace = true
 
 bootstrap-agent-client.workspace = true
 bootstrap-agent-lockstep-client.workspace = true
+bootstrap-agent-lockstep-types.workspace = true
 omicron-ddm-admin-client.workspace = true
 gateway-client.workspace = true
 gateway-types.workspace = true
